refactor: follow the file naming convention
