+++ src/app/products/[slug]/components/attribute-table.tsx
