+++ src/app/products/[slug]/components/carousel-section.tsx
