+++ src/app/products/[slug]/components/price-section.tsx
